pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cmux"
version = "0.1.0"
description = "PM Command Center for Parallel AI Sessions"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.9.0",
  "libtmux>=0.28.0",
  "pydantic>=2.0",
  "textual>=0.40.0",
  "rich>=13.0",
  "python-pptx>=0.6.21",
  "python-docx>=1.0",
  "httpx>=0.25.0",
  "pyyaml>=6.0",
  "eval_type_backport>=0.2.0; python_version < '3.10'",
]

[project.scripts]
cmux = "cmux.core.cli:app"

[tool.setuptools.packages.find]
include = ["cmux*"]
exclude = ["docs*", "cmux.egg-info*"]
